104.17 se corrige el filtro por distancia que estaba tomando null 113.4 task eliminar postulacion adopcion 113.5 task eliminar postulacion transito 106.12 mensajes, comento en el parser "usuarioPregunta" hasta que se modifique el servidor (no se pueden recibir los mensajes)

+++ app/src/main/java/com/tdp2grupo9/fragment/adopcion/ResultadosBusquedaFragment.java
@@ -77,7 +77,7 @@ public class ResultadosBusquedaFragment extends PublicacionesFragment {
         publicacion.setCompatibleCon(getCompatibleConPublicacion(arguments.getInt("compatiblecon", 0)));
         publicacion.setVacunasAlDia(getVacuntasAlDiaPublicacion(arguments.getInt("vacunas", 0)));
         publicacion.setPapelesAlDia(getPapelesAlDiaPublicacion(arguments.getInt("papeles", 0)));
-        publicacion.setDistancia(getDistanciaPublicacion(arguments.getDouble("distancia", -1.0)));
+        publicacion.setDistancia(getDistanciaPublicacion(arguments.getInt("distancia", -1)));
         publicacion.setLongitud(getLongitudPublicacion(arguments.getDouble("longitud", -1.0)));
         publicacion.setLatitud(getLatitudPublicacion(arguments.getDouble("latitud", -1.0)));
